Rename standard package to resource
